+++ dmwg_data_pyutils/__main__.py
@@ -5,7 +5,7 @@ import argparse
 import datetime
 import sys
 
-from dmwg_data_pyutils.logger import Logger
+from dmwg_data_pyutils.common.logger import Logger
 from dmwg_data_pyutils.subcommands import ParseNextStrain
